requirements.txt
pandas
textblob
google-generativeai
python-dotenv